[stdlib] Remove `Int` constructor from `StringSlice`
This has a dependency on `UInt` through its `base` argument, but we
cannot change it to `Int` because that would cause recursive definition
problems. Luckily, `Int` is already constructible from `IntableRaising`,
so string-like types should still convert to `Int`, and only cases where
a different `base` is specfied need to be changed (and since this
constructor is trivial, it should be easy to do this). Eventually, we
might reintroduce this, but for now, it's just an obstacle to SIMD
unification.

MODULAR_ORIG_COMMIT_REV_ID: b0404a175b322c8d84c08e2d36ab9832f94117cc