[bazel] Improve build perf on macOS
With recent bazel changes there is a significant difference in perf when
you have large directory
trees being used in every action. In this case the sysroots have
thousands of files,
and each were symlinked individually. Now instead they are symlinked
once as the top level
directory, and everything under that comes with it.

This affects linux as well but my experience testing was that the less
memory you have the
more of an issue it was, so testing on beefy remote Linux machines
doesn't show as much of a wall
time improvement.

As a follow up I want to do this with the toolchain files as well, but
there are 10x fewer than there
were with the sysroots, so it's likely less of an issue.

MODULAR_ORIG_COMMIT_REV_ID: feb450e6d02998eef26a417d99a850564d725d1a