[Driver] Use Nanobind to convert tensors to DLPack
We are currently using DLManagedTensor, from the 0.8 version of the
spec. This gives us free compatibility with the newer
DLManagedTensorVersioned, from 1.0 onwards.

It also handles some of the extra args like `max_version` for us.

Note on the modified test: Per the DLPack standard, we should transfer
if possible. With the new implementation, we are able to do so both to
and from numpy, so this test is no longer applicable. The test was
originally introduced to test a limitation of numpy.

MODULAR_ORIG_COMMIT_REV_ID: 8e1e25f7963ed3b3d49c24b9ee48f3fccb506575