[Pipelines] Add graphs to slice Eagle hidden states
When passing hidden states from the target to the draft model with Eagle
we need to extract the values for accepted tokens and also remove values
for tokens that are in sequences that have reached eos. Without these
slicing operations we are limited to running with batch size 1.

MODULAR_ORIG_COMMIT_REV_ID: 88899bd0e91732184a7932806f4e1bb08c29e575