[stdlib] Allows `DevicePassable` types to convert to other similar types.
The current `DevicePassable` trait is limited from mapping host type `A`
to a singular device type `B`. However, there are cases (specifically
when origins are in effect) that a type may want to be able to map to
_multiple_ different types. This is really only useful (now) for
`UnsafePointer` and `LayoutTensor` but this may extend to other types in
the future.

MODULAR_ORIG_COMMIT_REV_ID: b756bec97bcdfee06c47dda8360071a52a1eb713