[stdlib] polish: Remove `List[Scalar].extend(Span[Scalar])` overload
This overload of `List` was previously used to optimize extending a list
to a `memcpy()` in the case that the list element was a simple scalar
value.

However, as of a few months ago, the more general `List.extend(Span)`
overload provides that optimization by checking
`Self.T.__copyinit__is_trivial`, so this overload is no longer needed.

Additionally, this overload is ambiguous, leading to an ambiguity error
during the refactoring to drop `Copyable` as a requirement for `List`
elements.

MODULAR_ORIG_COMMIT_REV_ID: d55d8347fc160369d3c295d0b8ebe83904ac086f