[mojo-lang] Switch `__next__` to throwing StopIteration
This completes the migration of `Iterator` to Python-style design where
`__next__` raises at the end of iteration instead of having a separate
`__has_next__` method. In addition to being Pythonic, this is also
easier
to implement and use, more safe for clients, and can be slightly more
efficient in some cases.

Now that the mojo compiler's lowering of for-each loops can support
a raising `__next__` method in iterators, take the next step and switch
`Iterator` to requiring a raising next, and completely eliminate
`__has_next__` from Mojo.  Doing this requires reworking parameter
for lowering a bit, introducing a new `paramfor_has_next` helper,
but the biggest change is reworking all the testcases and iterators in
the tree.

I would like for this to be phase-in-able somehow, but I don't see a
practical way to do so.  We just have to accept this as a major breaking
change for folks that have custom iterators unfortunately.

MODULAR_ORIG_COMMIT_REV_ID: a3918eccc5760afc06d8ae19cbde6a61a5afb45a