[mojo-lang] Rewrite "init" result inference.
Now that parameter inference can infer parameters in arbitrary
order, we can remove some egregious (and flaky!) hacks for inferring
Self parameters in `__init__` methods that declare a custom "out self"
with computed parameter types.  We correctly infer these at the end
of parameter inference AFTER all the parameters they depend on are
resolved, instead of the previous fraught thing.

This should be a significant step forward in terms of correctness and
reliability.   This fixes MOCO-1826 and probably many other issues.

One thing it exposes is that the old code was incorrectly laundering
CompTimeOrigins to StaticConstant origins.  I'm not sure how this
ever worked, but this is correct. We get a new stdlib hack to enable
the old code until we can get rid of it.

MODULAR_ORIG_COMMIT_REV_ID: 0e70619c045e4538a87c4910348a783bce9f19c7